modify all existing layer classes to use add_param and to no longer define get_params themselves
